@@ -39,20 +39,11 @@ class Placeholder implements OutputElementInterface
         'height' => 10
     ];
 
-    /** @var string */
-    private $type;
-
-    /** @var string */
-    private $fileSvg;
-
-    /** @var string */
-    private $filePng;
-
-    /** @var int */
-    private $width;
-
-    /** @var int */
-    private $height;
+    private string $type;
+    private string $fileSvg;
+    private string $filePng;
+    private int $width;
+    private int $height;
 
     public static function create(array $type): self
     {
