@@ -14,11 +14,8 @@ class QrPaymentReferenceGenerator implements SelfValidatableInterface
 {
     use SelfValidatableTrait;
 
-    /** @var string */
-    private $customerIdentificationNumber;
-
-    /** @var string */
-    private $referenceNumber;
+    private ?string $customerIdentificationNumber = null;
+    private string $referenceNumber;
 
     public static function generate(?string $customerIdentificationNumber, string $referenceNumber): string
     {
