@@ -14,10 +14,9 @@ class RfCreditorReferenceGenerator implements SelfValidatableInterface
 {
     use SelfValidatableTrait;
 
-    /** @var string */
-    protected $reference;
+    private string $reference;
 
-    public static function generate(string $reference) : string
+    public static function generate(string $reference): string
     {
         $generator = new self($reference);
 
@@ -29,7 +28,7 @@ class RfCreditorReferenceGenerator implements SelfValidatableInterface
         $this->reference = StringModifier::stripWhitespace($reference);
     }
 
-    public function doGenerate() : string
+    public function doGenerate(): string
     {
         if (!$this->isValid()) {
             throw new InvalidCreditorReferenceException(
