@@ -13,12 +13,12 @@ class TranslationTest extends TestCase
     /**
      * @dataProvider allTranslationsProvider
      */
-    public function testAllByLanguage($locale, $subset)
+    public function testAllByLanguage(string $locale, array $subset): void
     {
         $this->assertArraySubset($subset, Translation::getAllByLanguage($locale));
     }
 
-    public function allTranslationsProvider()
+    public function allTranslationsProvider(): array
     {
         return [
             ['de', ['paymentPart' => 'Zahlteil']],
@@ -31,12 +31,12 @@ class TranslationTest extends TestCase
     /**
      * @dataProvider singleTranslationProvider
      */
-    public function testGet($locale, $key, $translation)
+    public function testGet(string $locale, string $key, string $translation)
     {
         $this->assertSame($translation, Translation::get($key, $locale));
     }
 
-    public function singleTranslationProvider()
+    public function singleTranslationProvider(): array
     {
         return [
             ['de', 'paymentPart', 'Zahlteil'],
